LPS-124494 Move TimeSpanSelector from Chart to Main component
