+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/Chart.js
@@ -28,7 +28,6 @@ import {
 import {
 	ChartDispatchContext,
 	ChartStateContext,
-	useDateTitle,
 	useIsPreviousPeriodButtonDisabled,
 } from '../context/ChartStateContext';
 import ConnectionContext from '../context/ConnectionContext';
@@ -37,7 +36,6 @@ import {generateDateFormatters as dateFormat} from '../utils/dateFormat';
 import {numberFormat} from '../utils/numberFormat';
 import {ActiveDot as CustomActiveDot, Dot as CustomDot} from './CustomDots';
 import CustomTooltip from './CustomTooltip';
-import TimeSpanSelector from './TimeSpanSelector';
 
 const CHART_COLORS = {
 	analyticsReportsHistoricalReads: '#50D2A0',
@@ -131,11 +129,7 @@ function legendFormatterGenerator(
 	};
 }
 
-export default function Chart({
-	dataProviders = [],
-	publishDate,
-	timeSpanOptions,
-}) {
+export default function Chart({dataProviders = [], publishDate}) {
 	const {validAnalyticsConnection} = useContext(ConnectionContext);
 
 	const storeDispatch = useContext(StoreDispatchContext);
@@ -144,23 +138,15 @@ export default function Chart({
 
 	const {languageTag, publishedToday} = useContext(StoreStateContext);
 
-	const {
-		dataSet,
-		loading,
-		timeRange,
-		timeSpanKey,
-		timeSpanOffset,
-	} = useContext(ChartStateContext);
-
-	const {firstDate, lastDate} = useDateTitle();
+	const {dataSet, loading, timeRange, timeSpanKey, timeSpanOffset} =
+		useContext(ChartStateContext);
 
 	const isPreviousPeriodButtonDisabled = useIsPreviousPeriodButtonDisabled();
 
-	const dateFormatters = useMemo(() => dateFormat(languageTag), [
-		languageTag,
-	]);
-
-	const title = dateFormatters.formatChartTitle([firstDate, lastDate]);
+	const dateFormatters = useMemo(
+		() => dateFormat(languageTag),
+		[languageTag]
+	);
 
 	const isMounted = useIsMounted();
 
@@ -196,8 +182,7 @@ export default function Chart({
 							...dataSetItems,
 							...data[i].value,
 						};
-					}
-					else {
+					} else {
 						storeDispatch({type: 'ADD_WARNING'});
 					}
 				}
@@ -212,8 +197,7 @@ export default function Chart({
 					validAnalyticsConnection,
 				});
 			});
-		}
-		else {
+		} else {
 			chartDispatch({
 				payload: {
 					keys,
@@ -260,19 +244,6 @@ export default function Chart({
 
 	return (
 		<>
-			{!!timeSpanOptions.length && (
-				<div className="c-mb-3 c-mt-4">
-					<TimeSpanSelector
-						disabledNextTimeSpan={timeSpanOffset === 0}
-						disabledPreviousPeriodButton={
-							isPreviousPeriodButtonDisabled
-						}
-						timeSpanKey={timeSpanKey}
-						timeSpanOptions={timeSpanOptions}
-					/>
-				</div>
-			)}
-
 			{dataSet ? (
 				<div className={lineChartWrapperClasses}>
 					{loading && (
@@ -281,8 +252,6 @@ export default function Chart({
 							small
 						/>
 					)}
-
-					{title && <h5 className="mb-3">{title}</h5>}
 
 					<div className="line-chart">
 						<LineChart
