+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/components/Main.js
@@ -10,12 +10,22 @@
  */
 
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useContext, useMemo} from 'react';
 
 import BasicInformation from './BasicInformation';
 import Chart from './Chart';
+import TimeSpanSelector from './TimeSpanSelector';
 import TotalCount from './TotalCount';
 import TrafficSources from './TrafficSources';
+
+import {
+	ChartStateContext,
+	useDateTitle,
+	useIsPreviousPeriodButtonDisabled,
+} from '../context/ChartStateContext';
+import {StoreStateContext} from '../context/StoreContext';
+
+import {generateDateFormatters as dateFormat} from '../utils/dateFormat';
 
 export default function Main({
 	author,
@@ -31,6 +41,20 @@ export default function Main({
 	trafficSourcesDataProvider,
 	viewURLs,
 }) {
+	const {timeSpanKey, timeSpanOffset} = useContext(ChartStateContext);
+	const {languageTag} = useContext(StoreStateContext);
+
+	const isPreviousPeriodButtonDisabled = useIsPreviousPeriodButtonDisabled();
+
+	const {firstDate, lastDate} = useDateTitle();
+
+	const dateFormatters = useMemo(
+		() => dateFormat(languageTag),
+		[languageTag]
+	);
+
+	const title = dateFormatters.formatChartTitle([firstDate, lastDate]);
+
 	return (
 		<div className="pb-3 px-3">
 			<BasicInformation
@@ -41,6 +65,20 @@ export default function Main({
 				title={pageTitle}
 				viewURLs={viewURLs}
 			/>
+
+			{!!timeSpanOptions.length && (
+				<div className="c-mb-2 c-mt-4">
+					<TimeSpanSelector
+						disabledNextTimeSpan={timeSpanOffset === 0}
+						disabledPreviousPeriodButton={
+							isPreviousPeriodButtonDisabled
+						}
+						timeSpanKey={timeSpanKey}
+						timeSpanOptions={timeSpanOptions}
+					/>
+				</div>
+			)}
+			{title && <h5 className="c-mb-4">{title}</h5>}
 
 			<h5 className="mt-3 sheet-subtitle">
 				{Liferay.Language.get('engagement')}
